Updates weapons system to be 32 bit

+++ c/Objects/Weapons.c
@@ -247,7 +247,7 @@ char Weapon_aimPosList[10][2] = {
 	
 	This way, I can avoid the OPs and save space. For debug, it will remain.
 */
-unsigned short Weapon_props[72] = {
+unsigned int Weapon_props[72] = {
     // row 1
         // jetpack
         0,
